+++ src/Responses/Responses/CreateStreamedResponse.php
@@ -29,8 +29,11 @@ use OpenAI\Responses\Responses\Streaming\OutputTextDone;
 use OpenAI\Responses\Responses\Streaming\ReasoningSummaryPart;
 use OpenAI\Responses\Responses\Streaming\ReasoningSummaryTextDelta;
 use OpenAI\Responses\Responses\Streaming\ReasoningSummaryTextDone;
+use OpenAI\Responses\Responses\Streaming\ReasoningTextDelta;
+use OpenAI\Responses\Responses\Streaming\ReasoningTextDone;
 use OpenAI\Responses\Responses\Streaming\RefusalDelta;
 use OpenAI\Responses\Responses\Streaming\RefusalDone;
+use OpenAI\Responses\Responses\Streaming\Response;
 use OpenAI\Responses\Responses\Streaming\WebSearchCall;
 use OpenAI\Testing\Responses\Concerns\FakeableForStreamedResponse;
 
@@ -50,7 +53,7 @@ final class CreateStreamedResponse implements ResponseContract
 
     private function __construct(
         public readonly string $event,
-        public readonly CreateResponse|OutputItem|ContentPart|OutputTextDelta|OutputTextAnnotationAdded|OutputTextDone|RefusalDelta|RefusalDone|FunctionCallArgumentsDelta|FunctionCallArgumentsDone|FileSearchCall|WebSearchCall|CodeInterpreterCall|CodeInterpreterCodeDelta|CodeInterpreterCodeDone|ReasoningSummaryPart|ReasoningSummaryTextDelta|ReasoningSummaryTextDone|McpListTools|McpListToolsInProgress|McpCall|McpCallArgumentsDelta|McpCallArgumentsDone|ImageGenerationPart|ImageGenerationPartialImage|Error $response,
+        public readonly Response|OutputItem|ContentPart|OutputTextDelta|OutputTextAnnotationAdded|OutputTextDone|RefusalDelta|RefusalDone|FunctionCallArgumentsDelta|FunctionCallArgumentsDone|FileSearchCall|WebSearchCall|CodeInterpreterCall|CodeInterpreterCodeDelta|CodeInterpreterCodeDone|ReasoningSummaryPart|ReasoningSummaryTextDelta|ReasoningSummaryTextDone|ReasoningTextDelta|ReasoningTextDone|McpListTools|McpListToolsInProgress|McpCall|McpCallArgumentsDelta|McpCallArgumentsDone|ImageGenerationPart|ImageGenerationPartialImage|Error $response,
     ) {}
 
     /**
@@ -64,10 +67,11 @@ final class CreateStreamedResponse implements ResponseContract
 
         $response = match ($event) {
             'response.created',
+            'response.queued',
             'response.in_progress',
             'response.completed',
             'response.failed',
-            'response.incomplete' => CreateResponse::from($attributes['response'], $meta), // @phpstan-ignore-line
+            'response.incomplete' => Response::from($attributes, $meta), // @phpstan-ignore-line
             'response.output_item.added',
             'response.output_item.done' => OutputItem::from($attributes, $meta), // @phpstan-ignore-line
             'response.content_part.added',
@@ -95,6 +99,8 @@ final class CreateStreamedResponse implements ResponseContract
             'response.reasoning_summary_part.done' => ReasoningSummaryPart::from($attributes, $meta), // @phpstan-ignore-line
             'response.reasoning_summary_text.delta' => ReasoningSummaryTextDelta::from($attributes, $meta), // @phpstan-ignore-line
             'response.reasoning_summary_text.done' => ReasoningSummaryTextDone::from($attributes, $meta), // @phpstan-ignore-line
+            'response.reasoning_text.delta' => ReasoningTextDelta::from($attributes, $meta), // @phpstan-ignore-line
+            'response.reasoning_text.done' => ReasoningTextDone::from($attributes, $meta), // @phpstan-ignore-line
             'response.mcp_list_tools.in_progress' => McpListToolsInProgress::from($attributes, $meta), // @phpstan-ignore-line
             'response.mcp_list_tools.failed',
             'response.mcp_list_tools.completed' => McpListTools::from($attributes, $meta), // @phpstan-ignore-line