+++ src/Responses/Responses/Streaming/OutputTextDelta.php
@@ -12,7 +12,7 @@ use OpenAI\Responses\Meta\MetaInformation;
 use OpenAI\Testing\Responses\Concerns\Fakeable;
 
 /**
- * @phpstan-type OutputTextType array{content_index: int, delta: string, item_id: string, output_index: int}
+ * @phpstan-type OutputTextType array{type: string, content_index: int, delta: string, item_id: string, output_index: int, sequence_number: int}
  *
  * @implements ResponseContract<OutputTextType>
  */
@@ -27,10 +27,12 @@ final class OutputTextDelta implements ResponseContract, ResponseHasMetaInformat
     use HasMetaInformation;
 
     private function __construct(
+        public readonly string $type,
         public readonly int $contentIndex,
         public readonly string $delta,
         public readonly string $itemId,
         public readonly int $outputIndex,
+        public readonly int $sequenceNumber,
         private readonly MetaInformation $meta,
     ) {}
 
@@ -40,10 +42,12 @@ final class OutputTextDelta implements ResponseContract, ResponseHasMetaInformat
     public static function from(array $attributes, MetaInformation $meta): self
     {
         return new self(
+            type: $attributes['type'],
             contentIndex: $attributes['content_index'],
             delta: $attributes['delta'],
             itemId: $attributes['item_id'],
             outputIndex: $attributes['output_index'],
+            sequenceNumber: $attributes['sequence_number'],
             meta: $meta,
         );
     }
@@ -54,10 +58,12 @@ final class OutputTextDelta implements ResponseContract, ResponseHasMetaInformat
     public function toArray(): array
     {
         return [
+            'type' => $this->type,
             'content_index' => $this->contentIndex,
             'delta' => $this->delta,
             'item_id' => $this->itemId,
             'output_index' => $this->outputIndex,
+            'sequence_number' => $this->sequenceNumber,
         ];
     }
 }