+++ src/Responses/StreamResponse.php
@@ -57,10 +57,11 @@ final class StreamResponse implements ResponseHasMetaInformationContract, Respon
             $response = json_decode($data, true, flags: JSON_THROW_ON_ERROR);
 
             if (isset($response['error'])) {
-                throw new ErrorException($response['error'], $this->response->getStatusCode());
+                throw new ErrorException($response['error'], $this->response);
             }
 
-            if (isset($response['type']) && $response['type'] === 'ping') {
+            $skippableTypes = ['ping', 'keepalive'];
+            if (isset($response['type']) && in_array($response['type'], $skippableTypes, true)) {
                 continue;
             }
 
@@ -95,7 +96,6 @@ final class StreamResponse implements ResponseHasMetaInformationContract, Respon
 
     public function meta(): MetaInformation
     {
-        // @phpstan-ignore-next-line
         return MetaInformation::from($this->response->getHeaders());
     }
 }